Rename variables for clarity Moved structureFromFile to FileUtils
